% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coord.R
\name{add_coord}
\alias{add_coord}
\title{A cleanAddresses Function}
\usage{
add_coord(street, city, state, zip, identifier)
}
\arguments{
\item{street}{The column name containing street name and number}

\item{city}{The column name containing city name}

\item{state}{The column name containing state name}

\item{zip}{The column name containing zip codes}

\item{identifier}{The column name containing unique record ids}
}
\description{
This function creates a simple address tibble that can be passed through censusxy() to get x and y coordinates for each record. Best used after cleaning the street field with cleanAddresses::simplify_street. Can be joined directly into a dataset.
}
\examples{
myData$newField <- simplify_street(myData$rawStreetField)
myCoordinates <- add_coord(street = myData$newField, city = myData$cityField, state = myData$stateField, zip = myData$zipField, identifier = myData$Id)
}
\keyword{address}
\keyword{coordinates,}
