% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_coord.R
\name{round_coord}
\alias{round_coord}
\title{A cleanAddresses Function}
\usage{
round_coord(lattitude, longitude, distance)
}
\arguments{
\item{lattitude}{The vector of Lattitude values (usually a column from a dataset)}

\item{longitude}{The vector of Longitude values (usually a column from a dataset)}

\item{distance}{The number of degree decimal places to round the coordinate. 1 degree, or zero decimal places, rounds to an accuracy of approximately 111km. Each additional decimal place is 10 times more accurate in distance.}
}
\description{
This function takes in 2 vectors of coordinate values - one for lattitude and one for longitude. The result is a 4-column tibble the same length and order as the input values. The new coordinates can be accessed in column 3 and 4.
}
\examples{
# To round coordinates to an accuracy of approximately 1.11km. Returns 4 columns (two original, and two rounded)
round_coord(lattitude = myData$lat, longitude = myData$long, distance = 2)
# To get the rounded lattitude only
round_coord(lattitude = myData$lat, longitude = myData$long, distance = 2)[,4]
 
}
\keyword{address}
\keyword{coordinates,}
