% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_street.R
\name{simplify_street}
\alias{simplify_street}
\title{A cleanAddresses Function}
\usage{
simplify_street(street)
}
\arguments{
\item{street}{takes in the vector of character street names (usually a column from a dataset)}
}
\description{
This function takes in a character vector of character address text. The entry data MUST start with street numbers but has the option to include or exclude City, State, and Zip fields
The result is a vector of the same length as the input data, made up of simplified street addresses only. Any City, State, or Zip will be cut off. The result can be applied directly as a new column in a dataset
}
\examples{
myData$newField <- simplify_street(myData$rawStreetField)
}
\keyword{address}
\keyword{street,}
